pyproject.toml
[tool.poetry]
name = "crypto-signals"
version = "0.1.0"
description = "Crypto Sentinel - Automated Crypto Trading System"
authors = ["Admin <admin@example.com>"]
packages = [{include = "crypto_signals", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
# Configuration Management
pydantic-settings = "^2.7.0"
python-dotenv = "^1.0.0"
# Alpaca Trading & Market Data
alpaca-py = "^0.38.0"
loguru = "^0.7.3"
# Google Cloud Services
google-cloud-firestore = "^2.22.0"
google-cloud-bigquery = "^3.27.0"
google-cloud-secret-manager = "^2.21.0"
# HTTP Client
requests = "^2.31.0"
websockets = "^13.1"
# Data Analysis
pandas = "^2.2.0"
pandas-ta-classic = {git = "https://github.com/xgboosted/pandas-ta-classic.git", branch = "main"}
typer = "^0.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.0,<9.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pre-commit (<4.0)",
    "ruff (>=0.4.0,<0.5.0)",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "visual: marks tests as visual integration tests (require TEST_DISCORD_WEBHOOK)",
]

[tool.ruff]
line-length = 90
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by ruff-format
]

[tool.ruff.lint.isort]
# profile = "black" is implied by Ruff's defaults and other settings

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
